requirements.txt
opencv-python==4.5.1.48
pyzbar==0.1.8
requests==2.25.1
qrcode==6.1
pillow==8.1.2
paho-mqtt==1.5.1
python-telegram-bot==13.4.1
pyserial==3.5
pandas~=1.2.3
numpy~=1.20.1
scikit-learn~=0.24.1
scipy~=1.6.2
schedule~=1.1.0
tqdm~=4.59.0
yagmail~=0.14.247